requirements.txt
torch>=1.7.0
torchvision>=0.8.1
matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
numpy
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0
gradio==3.35.2
ultralytics == 8.0.120
streamlit==1.49.1
timm>=1.0.19
sklearn>=1.7.2

# Documentation requirements
mkdocs-material
mkdocstrings[python]
mkdocs-mermaid2-plugin
mkdocs-include-markdown-plugin
mkdocs-awesome-pages-plugin
black

# RMBG-1.4 requirements
typing
scikit-image
huggingface_hub
transformers>=4.39.1